Working on the shapes and images for the experiment
Functions/functions.py
#display_img() >> charges one image from a particular path and puts it
in a white window.

Functions/creative_functions.py
#points() > only for type A images, creates the positions for each
shape (dots or crosses) depending on the condition and the collective
predicate

# parametters_shapes()>> depending on the condition and the type of
image, determinates the size of the figure,
#the ratio of the shapes and the letters, the width of the lines and
the positions

#shape_by_shape() >>
 for color,condition and shape, creates two figures: a triangle and a
square made of shapes. (type A)

#letter_with_connctions() >> for color,condition and shape, creates a
figure with a letter connected to number of shapes
(type B)

#random_letter() >> creates a figure with a letter without connections

#parametters() >> for complete images, positions, etc.

#image_A() and #image_B() create the complete pictures for the
experiment in the folder /Images